Remove kubectl dependency, validate k8s server version via api (#396)
* Remove kubectl dependency, validate k8s server version via api


* Remove unused MockKubectl


* Remame kubectl.go to version.go
