+++ cli/cmd/check_test.go
@@ -11,28 +11,6 @@ import (
 
 func TestCheckStatus(t *testing.T) {
 	t.Run("Prints expected output", func(t *testing.T) {
-		kubectl := &k8s.MockKubectl{}
-		kubectl.SelfCheckResultsToReturn = []*healthcheckPb.CheckResult{
-			{
-				SubsystemName:         k8s.KubectlSubsystemName,
-				CheckDescription:      k8s.KubectlConnectivityCheckDescription,
-				Status:                healthcheckPb.CheckStatus_OK,
-				FriendlyMessageToUser: "This shouldnt be printed",
-			},
-			{
-				SubsystemName:         k8s.KubectlSubsystemName,
-				CheckDescription:      k8s.KubectlIsInstalledCheckDescription,
-				Status:                healthcheckPb.CheckStatus_FAIL,
-				FriendlyMessageToUser: "This should contain instructions for fail",
-			},
-			{
-				SubsystemName:         k8s.KubectlSubsystemName,
-				CheckDescription:      k8s.KubectlVersionCheckDescription,
-				Status:                healthcheckPb.CheckStatus_ERROR,
-				FriendlyMessageToUser: "This should contain instructions for err",
-			},
-		}
-
 		kubeApi := &k8s.MockKubeApi{}
 		kubeApi.SelfCheckResultsToReturn = []*healthcheckPb.CheckResult{
 			{
@@ -45,12 +23,18 @@ func TestCheckStatus(t *testing.T) {
 				SubsystemName:         k8s.KubeapiSubsystemName,
 				CheckDescription:      k8s.KubeapiAccessCheckDescription,
 				Status:                healthcheckPb.CheckStatus_OK,
-				FriendlyMessageToUser: "This shouldnt be printed",
+				FriendlyMessageToUser: "This shouldn't be printed",
+			},
+			{
+				SubsystemName:         k8s.KubeapiSubsystemName,
+				CheckDescription:      k8s.KubeapiVersionCheckDescription,
+				Status:                healthcheckPb.CheckStatus_ERROR,
+				FriendlyMessageToUser: "This should contain instructions for err",
 			},
 		}
 
 		output := bytes.NewBufferString("")
-		checkStatus(output, kubectl, kubeApi)
+		checkStatus(output, kubeApi)
 
 		goldenFileBytes, err := ioutil.ReadFile("testdata/status_busy_output.golden")
 		if err != nil {
